# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/puntajes": {
      "get": {
        "operationId": "PuntajeController_listar",
        "summary": "API para obtener el listado de puntajes",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Puntaje"
        ]
      },
      "post": {
        "operationId": "PuntajeController_crear",
        "summary": "API para crear un nuevo puntaje",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo puntaje utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPuntajeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Puntaje"
        ]
      }
    },
...
    "/puntajes/{id}": {
      "get": {
        "operationId": "PuntajeController_obtenerPorId",
        "summary": "API para obtener un puntaje por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del puntaje que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Puntaje"
        ]
      },
      "patch": {
        "operationId": "PuntajeController_actualizar",
        "summary": "API para actualizar un puntaje",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del puntaje que deseas modificar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un puntaje existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarPuntajeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Puntaje"
        ]
      }
    },
    "/puntajes/eliminar/{id}": {
      "delete": {
        "operationId": "PuntajeController_eliminar",
        "summary": "API para eliminar un puntaje por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del puntaje que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Puntaje"
        ]
      }
    },
    "/puntajes/soft-delete/{id}": {
      "patch": {
        "operationId": "PuntajeController_softDelete",
        "summary": "API para eliminar un puntaje por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del puntaje que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Puntaje"
        ]
      }
    },
    "/puntajes/restore/{id}": {
      "patch": {
        "operationId": "PuntajeController_restore",
        "summary": "API para restaurar un puntaje eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del puntaje que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Puntaje"
        ]
      }
    },
    "/rondas": {
      "get": {
        "operationId": "RondaController_listar",
        "summary": "API para obtener el listado de rondas",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ronda"
        ]
      },
      "post": {
        "operationId": "RondaController_crear",
        "summary": "API para crear un nuevo ronda",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Esta API permite crear un nuevo ronda utilizando los datos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearRondaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ronda"
        ]
      }
    },
    "/rondas/{id}": {
      "get": {
        "operationId": "RondaController_obtenerPorId",
        "summary": "API para obtener un ronda por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del ronda que deseas obtener",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ronda"
        ]
      },
      "patch": {
        "operationId": "RondaController_actualizar",
        "summary": "API para actualizar un ronda",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del ronda que deseas modificar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Esta API permite actualizar un ronda existente utilizando los atributos proporcionados en el cuerpo de la solicitud.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarRondaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ronda"
        ]
      }
    },
    "/rondas/eliminar/{id}": {
      "delete": {
        "operationId": "RondaController_eliminar",
        "summary": "API para eliminar un ronda por su ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del ronda que deseas eliminar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ronda"
        ]
      }
    },
    "/rondas/soft-delete/{id}": {
      "patch": {
        "operationId": "RondaController_softDelete",
        "summary": "API para eliminar un ronda por su ID (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del ronda que deseas eliminar (soft delete)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ronda"
        ]
      }
    },
    "/rondas/restore/{id}": {
      "patch": {
        "operationId": "RondaController_restore",
        "summary": "API para restaurar un ronda eliminado (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del ronda que deseas restaurar",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ronda"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend",
    "description": "Documentación",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "CrearPuntajeDto": {
        "type": "object",
        "properties": {
          "puntaje": {
            "type": "number",
            "example": "50"
          },
          "rondaId": {
            "type": "number",
            "description": "ID de la ronda (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "equipoId": {
            "type": "number",
            "description": "ID del equipo (foreign key)",
            "format": "int64",
            "example": "1"
          }
        },
        "required": [
          "puntaje",
          "rondaId",
          "equipoId"
        ]
      },
      "ActualizarPuntajeDto": {
        "type": "object",
        "properties": {
          "puntaje": {
            "type": "number",
            "example": "50"
          },
          "ronda": {
            "type": "number",
            "description": "ID de la ronda (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "equipo": {
            "type": "number",
            "description": "ID del equipo (foreign key)",
            "format": "int64",
            "example": "1"
          }
        },
        "required": [
          "puntaje",
          "ronda",
          "equipo"
        ]
      },
      "CrearRondaDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "example": "pendiente | en_curso | finalizada"
          },
          "pistaId": {
            "type": "number",
            "description": "ID de la pista (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "equipoRojoId": {
            "type": "number",
            "description": "ID del equipo rojo (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "equipoAzulId": {
            "type": "number",
            "description": "ID del equipo azul (foreign key)",
            "format": "int64",
            "example": "1"
          }
        },
        "required": [
          "estado",
          "pistaId",
          "equipoRojoId",
          "equipoAzulId"
        ]
      },
      "ActualizarRondaDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "example": "pendiente | en_curso | finalizada"
          },
          "pista": {
            "type": "number",
            "description": "ID de la pista (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "equipo_rojo": {
            "type": "number",
            "description": "ID del equipo rojo (foreign key)",
            "format": "int64",
            "example": "1"
          },
          "equipo_azul": {
            "type": "number",
            "description": "ID del equipo azul (foreign key)",
            "format": "int64",
            "example": "1"
          }
        },
        "required": [
          "estado",
          "pista",
          "equipo_rojo",
          "equipo_azul"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}